Fix updating index block after adding new data blocks
When the size of a file increases by a write, ouichefs_file_get_block
will be used to allocate a new block for the file, and the block number
is written into the ouichefs_file_index_block index of the file.
After modifying the index, it must be marked as dirty for the change
to have a persistent effect. This commit adds the missing
mark_buffer_dirty call after modifying the index.

To reproduce the issue, create a new test.img using mkfs.ouichefs
and create a file backed by one block of data:
$ dd if=/dev/random of=mount/testfile bs=4096 count=1
This file is backed by the file index stored in block 256 and the
data block 257. Now sync to ensure the file is written to disk and
the index is unmarked as dirty. If we then enlarge the file such
that it requires to be backed by two data blocks, block 258 will be
used to store the data and 258 is added to the second slot of the
index. block_write_begin will take care of marking the newly
allocated data block 258 as dirty, but is unaware of the ouichefs
specific index. Thus, the index modification will be lost after
unmounting the filesystem:
$ sync
$ dd if=/dev/random of=mount/testfile skip=4096 bs=8192 count=1
$ umount mount && mount test.img mount
$ hexdump mount/testfile
[... omitted]
0001000 0000 0000 0000 0000 0000 0000 0000 0000
*
0002000

As we can observe, the whole second block of the file is filled with
zeros after remounting the filesystem.

Signed-off-by: Arian Kulmer <arian.kulmer@rwth-aachen.de>