Fix storing ouichefs index block on open
When opening an existing file with truncate it is
meant to be wiped. However if no actual write happens,
the index block update won't be saved.
Consider this c program as an example:

\#include <stdio.h>

int main(int argc, char *argv[]) {

    FILE *file = fopen(argv[1], "w"); // Open with truncate and write mode

    if (file == NULL) {
        perror("Error opening file");
        return 1;
    }

    fclose(file); // Close the file
    return 0;
}

When running the following commands on a fresh ouichefs image
we can observe the bug:
$ echo "some data" > mount/testfile
$ sync
$ ./a.out mount/testfile
$ umount mount

After this execution the index block of testfile still
points to the old data block.

Signed-off-by: Arian Kulmer <arian.kulmer@rwth-aachen.de>