add missing endianness conversions
Add missing endianness conversions throughout the file system implementation,
which should make it possible to use the file system on systems with
big endian as well.

For the ifree and bfree bitmaps, the conversions is slightly more complex,
because we must consider that the data type used for bitmaps in Linux
(unsigned long) could be only 32 bits large on some architectures.