project: Added reference to the next free sliced block in ouichefs super block and proper initialization of the root directory
As we want to maintain a list of partially filled blocks, we add a reference to the first partially filled block
in this list in our struct ouichefs_superblock (s_free_sliced_blocks) and initialize it to 0.
Further, when printing the superblock information, we now also print the information about a sliced block.
In the case of a file, the index_block in ouichefs_inode now either references the block with the list
of the blocks for this file or the new slice information for small files,
i.e. the slice number (5 bits) and the block number (27 bits).
We also want to properly initialize the root directory of the filesystem (in write_data_blocks()).
It creates and writes the root directory block to disk during the filesystem formatting.
We set the current and parent directory and everything else to zero.
After all, the root inode's index_block field points to this block, with only containing the files . and ..

Signed-off-by: David Ginten <david.ginten@rwth-aachen.de>