Use sysfs to check the state of the file system
In the pseudo file system sysfs, multiple variables (see project description) can be monitored to get a view of the current state of the FS.

fs.c:
As multiple partitions are concurrently needed, multiple partitions, each having its own sb, can be created.
They are differentiated by the name.

sysfs.c:
In order to manage multiple superblocks, a hash table is used for fast access.
Further, for each value, a show function exists that first collects the latest statistics and then prints the relevant data.
The collect stats function first collects the basic data from the sb_info and then calculates the rest of the data
with the help of the basic data in the superblock info, the inode list, and the list of partially filled blocks.

Signed-off-by: David Ginten <david.ginten@rwth-aachen.de>