ioctl command to display the content of a block.
An ioctl command has been implemented to display the content of a block. That means 32 slices (rows) each having 128 characters (128 Bytes).
The ioctl handler reads a whole sliced block (The block where the file specified by the file descriptor is stored)
from the kernel and uses a struct that can hold those 4096 bytes. This command only reads data from the kernel.

Then we have a user space file that calls this ioctl syscall (with the given fd and the struct where we expect the data to be written to).
It then prints each slice in a row, where a Byte is represented by a char. As the bitmap is more visible when represented
in bits, the 32 bits are also printed at the beginning. This is also true for the reference integer to the next block in the list of
partially filled blocks. It is also printed at the beginning as an int. (Not build into the module of course, just for review).

Signed-off-by: David Ginten <david.ginten@rwth-aachen.de>